document-converter/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mammoth==1.6.0
pypandoc==1.12
python-docx==1.1.0
Pillow==10.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
aiofiles==23.2.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-magic==0.4.27
lxml==5.1.0
odfpy==1.4.1